--- a/app/src/main/java/fr/neamar/kiss/DataHandler.java
+++ b/app/src/main/java/fr/neamar/kiss/DataHandler.java
@@ -418,7 +418,8 @@ public class DataHandler extends BroadcastReceiver
 		StringBuilder excluded = new StringBuilder();
 		for(String excludedItem : excludedList) {
 			if(!user.hasStringUserSuffix(excludedItem, '#')) {
-				excluded.append(excludedItem + ";");
+				excluded.append( excludedItem )
+                        .append( ";" );
 			}
 		}