--- a/app/src/main/java/fr/neamar/kiss/MainActivity.java
+++ b/app/src/main/java/fr/neamar/kiss/MainActivity.java
@@ -66,12 +66,13 @@ import fr.neamar.kiss.searcher.NullSearcher;
 import fr.neamar.kiss.searcher.QueryInterface;
 import fr.neamar.kiss.searcher.QuerySearcher;
 import fr.neamar.kiss.searcher.Searcher;
+import fr.neamar.kiss.ui.AnimatedListView;
 import fr.neamar.kiss.ui.BlockableListView;
 import fr.neamar.kiss.ui.BottomPullEffectView;
 import fr.neamar.kiss.ui.KeyboardScrollHider;
 import fr.neamar.kiss.utils.PackageManagerUtils;
 
-public class MainActivity extends Activity implements QueryInterface, KeyboardScrollHider.KeyboardHandler, View.OnTouchListener {
+public class MainActivity extends Activity implements QueryInterface, KeyboardScrollHider.KeyboardHandler, View.OnTouchListener, Searcher.DataObserver {
 
     public static final String START_LOAD = "fr.neamar.summon.START_LOAD";
     public static final String LOAD_OVER = "fr.neamar.summon.LOAD_OVER";
@@ -1142,5 +1143,19 @@ public class MainActivity extends Activity implements QueryInterface, KeyboardSc
         for (int appWidgetId : widgetIds.values()){
             addWidgetToLauncher(appWidgetId);
         }
+    }
+
+    @Override
+    public void beforeChange()
+    {
+        AnimatedListView listView = (AnimatedListView)this.list;
+        listView.prepareChangeAnim();
+    }
+
+    @Override
+    public void afterChange()
+    {
+        AnimatedListView listView = (AnimatedListView)this.list;
+        listView.animateChange();
     }
 }