--- a/app/src/main/java/fr/neamar/kiss/MainActivity.java
+++ b/app/src/main/java/fr/neamar/kiss/MainActivity.java
@@ -748,7 +748,7 @@ public class MainActivity extends Activity implements QueryInterface, KeyboardSc
                     //if list is empty
                     if ((this.list.getAdapter() == null) || (this.list.getAdapter().getCount() == 0)) {
                         searcher = new HistorySearcher(MainActivity.this);
-                        searcher.execute();
+                        searcher.executeOnExecutor( Searcher.SEARCH_THREAD );
                     }
                 }
             }
@@ -808,7 +808,7 @@ public class MainActivity extends Activity implements QueryInterface, KeyboardSc
         }
     }
 
-    private void displayLoader(Boolean display) {
+    public void displayLoader(Boolean display) {
         final View loaderBar = findViewById(R.id.loaderBar);
         final View launcherButton = findViewById(R.id.launcherButton);
 
@@ -861,7 +861,7 @@ public class MainActivity extends Activity implements QueryInterface, KeyboardSc
                 searcher.cancel(true);
             }
             searcher = new ApplicationsSearcher(MainActivity.this);
-            searcher.execute();
+            searcher.executeOnExecutor( Searcher.SEARCH_THREAD );
 
             // Reveal the bar
             if (android.os.Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
@@ -997,7 +997,7 @@ public class MainActivity extends Activity implements QueryInterface, KeyboardSc
         } else {
             searcher = new QuerySearcher(this, query);
         }
-        searcher.execute();
+        searcher.executeOnExecutor( Searcher.SEARCH_THREAD );
         displayQuickFavoritesBar(false, false);
     }