--- a/app/src/main/java/fr/neamar/kiss/adapter/RecordAdapter.java
+++ b/app/src/main/java/fr/neamar/kiss/adapter/RecordAdapter.java
@@ -2,6 +2,7 @@ package fr.neamar.kiss.adapter;
 
 import android.content.Context;
 import android.os.Handler;
+import android.util.Log;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
@@ -75,7 +76,21 @@ public class RecordAdapter extends ArrayAdapter<Result> {
 
     @Override
     public View getView(int position, View convertView, ViewGroup parent) {
-        return results.get(position).display(getContext(), results.size() - position, convertView);
+        if ( convertView != null )
+        {
+            if ( !(convertView.getTag() instanceof Integer) )
+                convertView = null;
+            else if ( (Integer)convertView.getTag() != getItemViewType( position ) )
+            {
+                // This is happening on HTC Desire X (Android 4.1.1, API 16)
+                //throw new IllegalStateException( "can't convert view from different type" );
+                convertView = null;
+            }
+        }
+        View view = results.get(position).display(getContext(), results.size() - position, convertView);
+        //Log.d( "TBog", "getView pos " + position + " convertView " + ((convertView == null) ? "null" : convertView.toString()) + " will return " + view.toString() );
+        view.setTag( getItemViewType( position ) );
+        return view;
     }
 
     public void onLongClick(final int pos, View v) {