… app/src/main/java/fr/neamar/kiss/dataprovider/ContactsProvider.java
… app/src/main/java/fr/neamar/kiss/dataprovider/PhoneProvider.java
… app/src/main/java/fr/neamar/kiss/dataprovider/Provider.java
… app/src/main/java/fr/neamar/kiss/dataprovider/SearchProvider.java
… app/src/main/java/fr/neamar/kiss/dataprovider/SettingsProvider.java
… app/src/main/java/fr/neamar/kiss/dataprovider/ShortcutsProvider.java
… app/src/main/java/fr/neamar/kiss/dataprovider/TogglesProvider.java
… app/src/main/java/fr/neamar/kiss/loader/LoadContactsPojos.java
… app/src/main/java/fr/neamar/kiss/loader/LoadSearchPojos.java
… app/src/main/java/fr/neamar/kiss/loader/LoadSettingsPojos.java
… app/src/main/java/fr/neamar/kiss/loader/LoadTogglesPojos.java
… app/src/main/java/fr/neamar/kiss/normalizer/StringNormalizer.java
… app/src/main/java/fr/neamar/kiss/pojo/AppPojo.java
… app/src/main/java/fr/neamar/kiss/pojo/ContactsPojo.java
… app/src/main/java/fr/neamar/kiss/pojo/Pojo.java
… app/src/main/java/fr/neamar/kiss/pojo/PojoComparator.java
… app/src/main/java/fr/neamar/kiss/pojo/PojoWithTags.java
… app/src/main/java/fr/neamar/kiss/result/AppResult.java
… app/src/main/java/fr/neamar/kiss/result/ContactsResult.java
… app/src/main/java/fr/neamar/kiss/result/Result.java
… app/src/main/java/fr/neamar/kiss/result/SearchResult.java
… app/src/main/java/fr/neamar/kiss/searcher/ApplicationsSearcher.java
… app/src/main/java/fr/neamar/kiss/searcher/HistorySearcher.java
… app/src/main/java/fr/neamar/kiss/searcher/NullSearcher.java
… app/src/main/java/fr/neamar/kiss/searcher/QueryInterface.java
… app/src/main/java/fr/neamar/kiss/searcher/QuerySearcher.java
… app/src/main/java/fr/neamar/kiss/searcher/Searcher.java
… app/src/main/java/fr/neamar/kiss/ui/AnimatedListView.java
… app/src/main/java/fr/neamar/kiss/utils/DamerauLevenshteinAlgorithm.java
… app/src/main/java/fr/neamar/kiss/utils/FuzzyScore.java