new algorithm
Timing: Time to run query `t` to completion: 163ms
Timing: Time to run query `<null>` to completion: 13ms
Timing: Time to run query `t` to completion: 51ms
Timing: Time to run query `<null>` to completion: 24ms
Timing: Time to run query `y` to completion: 52ms
Timing: Time to run query `yt` to completion: 22ms
Timing: Time to run query `ytu` to completion: 50ms
Timing: Time to run query `yt` to completion: 43ms
Timing: Time to run query `ytu` to completion: 57ms
Timing: Time to run query `yt` to completion: 41ms
Timing: Time to run query `ytu` to completion: 55ms
Timing: Time to run query `yt` to completion: 43ms
Timing: Time to run query `ytu` to completion: 43ms
Timing: Time to run query `yt` to completion: 34ms
Timing: Time to run query `ytu` to completion: 44ms
Timing: Time to run query `<null>` to completion: 14ms

old algorithm
Timing: Time to run query `t` to completion: 78ms
Timing: Time to run query `<null>` to completion: 7ms
Timing: Time to run query `t` to completion: 20ms
Timing: Time to run query `<null>` to completion: 7ms
Timing: Time to run query `y` to completion: 21ms
Timing: Time to run query `yt` to completion: 13ms
Timing: Time to run query `ytu` to completion: 27ms
Timing: Time to run query `yt` to completion: 27ms
Timing: Time to run query `ytu` to completion: 17ms
Timing: Time to run query `yt` to completion: 12ms
Timing: Time to run query `ytu` to completion: 15ms
Timing: Time to run query `yt` to completion: 16ms
Timing: Time to run query `ytu` to completion: 20ms
Timing: Time to run query `yt` to completion: 35ms
Timing: Time to run query `ytu` to completion: 33ms
Timing: Time to run query `<null>` to completion: 4ms