@Neamar 2018-01-14
Just tested,

Animations are great!
I did notice that the app got noticeably slower after the update, which is OK. However, I think the relevance for apps should be increased compared to, for instance, contacts alias. I'll have a look at the scores one of those days :)