@TBog 2018-01-25
@Neamar I had some problems with apps that had emoticons in the name and fixed it with commit [67a93be](https://github.com/Neamar/KISS/pull/766/commits/67a93bebd25ec904fad8cba2be62476714b765b8)
I can do the same for the main branch if you think it will get merged before this PR