@Neamar 2018-01-25
@TBog can you cherry-pick this in a new  PR? I'll merge it directly.

This PR (#766) is looking good, but probably require some time in the beta
channel, whereas the bugfix can go to prod now.

Thanks!

On Thu, Jan 25, 2018 at 8:07 AM TBog <notifications@github.com> wrote:

> @Neamar <https://github.com/neamar> I had some problems with apps that
> had emoticons in the name and fixed it with commit 67a93be
> <https://github.com/Neamar/KISS/pull/766/commits/67a93bebd25ec904fad8cba2be62476714b765b8>
> I can do the same for the main branch if you think it will get merged
> before this PR
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/Neamar/KISS/pull/766#issuecomment-360513594>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/AAgxDJ49Pz-nw8J4Kld69H3cwIMXfEQfks5tOKavgaJpZM4RAyJh>
> .
>