@Neamar 2018-01-29
K, I've been using it for some time now.

It works well.

It is slower than the current algorithm, but it also feels better. I'm going to merge, and we'll keep this in beta for some time to make sure everything is covered.

There were some conflicts, some of them related to the recent cherry-picking for emojis. I'm not sure what is the correct version if you can have a look? I kept the version on this branch.